nextQuestionCheck function is in new separate file
